Add support for DriverID to event data model
In order to support journeys being taken over by drivers, we need to
encode the DriverID alongside the generated Event record. This can
occur either in cases where a journey is paused and a driver physically
changed (where the driver is a natural person), or in the case of an
autonomous driving system in which the vehicle may take over driving
functions at certain automation levels.