Add an optional omitNull parameter to event model json encoding
By default, null values are excluded in the JSON map, but there are
times where this is undesirable and the full set of keys must be made
accessible (e.g. for preparing CSV columns). As flutter does not
support direct reflection of the class itself, this can be used as a
workaround when all class members need to be identified.